extracted date and weather parsing logic to date-temp-util
